requirements.txt
torch
transformers
deep-translator
SpeechRecognition
gunicorn
flask
flask-cors
python-dotenv
requests
bitsandbytes